requirements.txt
pytest==3.6.1
numpy==1.14.3
matplotlib==2.2.3
pyserial==3.4
scikit-image
